+++ src/commonmark/editor.ts
@@ -6,7 +6,6 @@ import { IExternalPluginProvider } from "../shared/editor-plugin";
 import { log } from "../shared/logger";
 import { createMenuPlugin } from "../shared/menu/plugin";
 import { createPreviewPlugin } from "./plugins/preview";
-import { commonmarkCodePasteHandler } from "../shared/prosemirror-plugins/code-paste-handler";
 import {
     commonmarkImageUpload,
     defaultImageUploadHandler,
@@ -116,7 +115,6 @@ export class CommonmarkEditor extends BaseView {
                         placeholderPlugin(this.options.placeholderText),
                         readonlyPlugin(),
                         tripleClickHandler,
-                        commonmarkCodePasteHandler,
                         textCopyHandlerPlugin,
                         ...pluginProvider.plugins.commonmark,
                     ],